+++ src/TowerDefense/TowerDefense/Tower.cs
@@ -9,7 +9,7 @@ namespace TowerDefense
     class Tower
     {
         protected virtual int Range { get; } = 1;
-        protected virtual int Power { get; } = 1;
+        protected virtual double Power { get; } = 1;
 		protected virtual double Accuracy { get; } = .75;
 
         private static readonly Random _random = new Random();
